+++ packages/moonstone/VideoPlayer/MediaControls.js
@@ -9,15 +9,6 @@ import IconButton from '../IconButton';
 
 import css from './VideoPlayer.less';
 
-const buttonLabels = {
-	previousLabel: 'Previous',
-	rewindLabel: 'Rewind',
-	fastForwardLabel: 'Fast Forward',
-	nextLabel: 'Next',
-	backLabel: 'Back',
-	moreLabel: 'More'
-};
-
 const Container = SpotlightContainerDecorator({enterTo: ''}, 'div');
 const MediaButton = onlyUpdateForKeys([
 	'children',
@@ -156,8 +147,6 @@ const MediaControls = kind({
 		forwardIcon: 'forward',
 		jumpBackwardIcon: 'skipbackward',
 		jumpForwardIcon: 'skipforward',
-		moreButtonCloseLabel: buttonLabels.backLabel,
-		moreButtonLabel: buttonLabels.moreLabel,
 		pauseIcon: 'pause',
 		playIcon: 'play'
 	},
@@ -173,7 +162,13 @@ const MediaControls = kind({
 			more: showMoreComponents
 		}),
 		moreIcon: ({showMoreComponents}) => showMoreComponents ? 'arrowshrinkleft' : 'ellipsis',
-		moreIconLabel: ({moreButtonCloseLabel, moreButtonLabel, showMoreComponents}) => showMoreComponents ? moreButtonCloseLabel : moreButtonLabel,
+		moreIconLabel: ({moreButtonCloseLabel, moreButtonLabel, showMoreComponents}) => {
+			if (showMoreComponents) {
+				return moreButtonCloseLabel != null ? moreButtonCloseLabel : $L('Back');
+			} else {
+				return moreButtonLabel != null ? moreButtonLabel : $L('More');
+			}
+		},
 		playPauseIcon: ({paused, pauseIcon, playIcon}) => (paused ? playIcon : pauseIcon),
 		playPauseLabel: ({paused, pauseLabel, playLabel}) => (paused ? playLabel : pauseLabel)
 	},
@@ -208,13 +203,6 @@ const MediaControls = kind({
 			...rest
 		} = props;
 
-		const {
-			previousLabel,
-			rewindLabel,
-			fastForwardLabel,
-			nextLabel
-		} = buttonLabels;
-
 		delete rest.moreButtonCloseLabel;
 		delete rest.moreButtonLabel;
 		delete rest.pauseIcon;
@@ -230,11 +218,11 @@ const MediaControls = kind({
 				<div className={css.centerComponentsContainer}>
 					<div className={centerClassName}>
 						<Container className={css.mediaControls} spotlightDisabled={!moreDisabled}> {/* rtl={false} */}
-							{noJumpButtons ? null : <MediaButton aria-label={$L(previousLabel)} backgroundOpacity="translucent" disabled={mediaDisabled || jumpButtonsDisabled} onClick={onJumpBackwardButtonClick}>{jumpBackwardIcon}</MediaButton>}
-							{noRateButtons ? null : <MediaButton aria-label={$L(rewindLabel)} backgroundOpacity="translucent" disabled={mediaDisabled || rateButtonsDisabled} onClick={onBackwardButtonClick}>{backwardIcon}</MediaButton>}
-							<MediaButton aria-label={$L(playPauseLabel)} className={spotlightDefaultClass} backgroundOpacity="translucent" onClick={onPlayButtonClick}>{playPauseIcon}</MediaButton>
-							{noRateButtons ? null : <MediaButton aria-label={$L(fastForwardLabel)} backgroundOpacity="translucent" disabled={mediaDisabled || rateButtonsDisabled} onClick={onForwardButtonClick}>{forwardIcon}</MediaButton>}
-							{noJumpButtons ? null : <MediaButton aria-label={$L(nextLabel)} backgroundOpacity="translucent" disabled={mediaDisabled || jumpButtonsDisabled} onClick={onJumpForwardButtonClick}>{jumpForwardIcon}</MediaButton>}
+							{noJumpButtons ? null : <MediaButton aria-label={$L('Previous')} backgroundOpacity="translucent" disabled={mediaDisabled || jumpButtonsDisabled} onClick={onJumpBackwardButtonClick}>{jumpBackwardIcon}</MediaButton>}
+							{noRateButtons ? null : <MediaButton aria-label={$L('Rewind')} backgroundOpacity="translucent" disabled={mediaDisabled || rateButtonsDisabled} onClick={onBackwardButtonClick}>{backwardIcon}</MediaButton>}
+							<MediaButton aria-label={playPauseLabel} className={spotlightDefaultClass} backgroundOpacity="translucent" onClick={onPlayButtonClick}>{playPauseIcon}</MediaButton>
+							{noRateButtons ? null : <MediaButton aria-label={$L('Fast Forward')} backgroundOpacity="translucent" disabled={mediaDisabled || rateButtonsDisabled} onClick={onForwardButtonClick}>{forwardIcon}</MediaButton>}
+							{noJumpButtons ? null : <MediaButton aria-label={$L('Next')} backgroundOpacity="translucent" disabled={mediaDisabled || jumpButtonsDisabled} onClick={onJumpForwardButtonClick}>{jumpForwardIcon}</MediaButton>}
 						</Container>
 						<Container className={css.moreControls} spotlightDisabled={moreDisabled}>
 							{children}
@@ -243,7 +231,19 @@ const MediaControls = kind({
 				</div>
 				<div className={css.rightComponents}>
 					{rightComponents}
-					{React.Children.count(children) ? <MediaButton aria-label={$L(moreIconLabel)} backgroundOpacity="translucent" className={css.moreButton} disabled={moreButtonDisabled} onClick={onToggleMore} tooltipText={$L(moreIconLabel)}>{moreIcon}</MediaButton> : null}
+					{React.Children.count(children) ? (
+						<MediaButton
+							aria-label={moreIconLabel}
+							backgroundOpacity="translucent"
+							className={css.moreButton}
+							disabled={moreButtonDisabled}
+							onClick={onToggleMore}
+							tooltipProps={{role: 'dialog'}}
+							tooltipText={moreIconLabel}
+						>
+							{moreIcon}
+						</MediaButton>
+					) : null}
 				</div>
 			</div>
 		);