+++ packages/ui/CHANGELOG.md
@@ -12,7 +12,15 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ### Fixed
 
-- `ui/PlaceholderDecorator` to update bounds of `Scroller` when the `visible` state changed.
+## [1.6.1] - 2017-08-07
+
+No significant changes.
+
+## [1.6.0] - 2017-08-04
+
+### Fixed
+
+- `ui/PlaceholderDecorator` to update bounds of `Scroller` when the `visible` state changed
 
 ## [1.5.0] - 2017-07-19